chore: use npm to install repositories instead of git modules
