+++ package.json
@@ -19,7 +19,7 @@
     "format": "prettier --write .",
     "lint": "eslint . --fix",
     "pack.exe": "shx rm -rf ./dist/tsconfig.tsbuildinfo && node ./building/scripts/pack-exe.js",
-    "prepare": "git submodule update --init --recursive && npm run build",
+    "prepare": "npm run build",
     "start.docker": "docker run -t setup_cpp .",
     "test": "run-p test.format test.lint test.tsc test.unit",
     "test.format": "prettier . --check",
@@ -38,7 +38,9 @@
     "mri": "^1.2.0",
     "semver": "^7.3.5",
     "untildify": "^4.0.0",
-    "which": "^2.0.2"
+    "which": "^2.0.2",
+    "setup-python": "https://github.com/actions/setup-python",
+    "msvc-dev-cmd": " https://github.com/ilammy/msvc-dev-cmd"
   },
   "devDependencies": {
     "@types/cross-spawn": "^6.0.2",
