+++ src/msvc/msvc.ts
@@ -2,24 +2,9 @@ import { setupChocoPack } from "../utils/setup/setupChocoPack"
 import { error, exportVariable, warning } from "@actions/core"
 import { existsSync } from "fs"
 import { isCI } from "../utils/env/isci"
+import { activateMSVC } from "../vcvarsall/vcvarsall"
 
 type MSVCVersion = "2015" | "2017" | "2019" | string
-
-function getArch(arch: string): string {
-  switch (arch) {
-    case "x32":
-    case "32":
-    case "ia32": {
-      return "x86"
-    }
-    case "64": {
-      return "x64"
-    }
-    default: {
-      return arch
-    }
-  }
-}
 
 export async function setupMSVC(
   version: MSVCVersion,
@@ -72,22 +57,4 @@ export async function setupMSVC(
   }
   // run vcvarsall.bat environment variables
   activateMSVC(VCTargetsPath, arch, toolset, sdk, uwp, spectre)
-}
-
-function activateMSVC(
-  VCTargetsPath: string | undefined,
-  arch: string,
-  toolset: string | undefined,
-  sdk?: string,
-  uwp?: boolean,
-  spectre?: boolean
-) {
-  if (VCTargetsPath !== undefined && existsSync(VCTargetsPath)) {
-    exportVariable("VCTargetsPath", VCTargetsPath)
-  }
-
-  // lazy load the action so it is not executed
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const { setupMSVCDevCmd } = require("./msvc-dev-cmd/index")
-  setupMSVCDevCmd(getArch(arch), sdk, toolset, uwp, spectre)
 }
